Functions created (User)
When a User created in the Firestore one function will add an empty wishlist to it and another one will create a Storage folder with the default profile picture. When a User is deleted from Auth it's Storage folder, Firestore Wishlist and User entry is deleted. When a Group is created in Firestore a function will add it's ID to it automatically.